Brushup Change -Added Functionality Descriptions To Files -Slight change to HealthReplenish.cs -Moved some scripts into scripts folder
