+++ Assets/Scripts/PlayerHealth.cs
@@ -2,6 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
+//Basic Health Script (used for oil) this is most likely a temporary script for testing 
 public class PlayerHealth : MonoBehaviour
 {
     // Player's health
